Add dto and refactoring
