+++ src/main/java/com/mydeveloperplanet/myawscdplanet/controller/HelloController.java
@@ -1,4 +1,4 @@
-package com.mydeveloperplanet.myawscdplanet;
+package com.mydeveloperplanet.myawscdplanet.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
