+++ src/main/java/com/mydeveloperplanet/myawscdplanet/dto/UserDto.java
@@ -1,0 +1,28 @@
+package com.mydeveloperplanet.myawscdplanet.dto;
+
+public class UserDto {
+
+  private String firstName;
+  private String lastName;
+
+  public UserDto(String firstName, String lastName) {
+    this.firstName = firstName;
+    this.lastName = lastName;
+  }
+
+  public String getFirstName() {
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public String getLastName() {
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+}
